fixing create delete list tests
